requirements.txt
openai
playwright
bs4
python-dotenv